refactor: break up useGameView composable
