+++ Dashboard/src/components/game/AuthModal.vue
@@ -1,52 +1,38 @@
 <script setup>
+import { useGameViewInject } from "../../composables/game/injectKey";
 import Dialog from "primevue/dialog";
 import Password from "primevue/password";
 import Button from "primevue/button";
 import Message from "primevue/message";
 
-const props = defineProps({
-  visible: Boolean,
-  authProfileId: [String, Number],
-  authPassphrase: String,
-  authLoading: Boolean,
-  authError: String,
-});
-
-const emit = defineEmits([
-  "update:visible",
-  "update:authPassphrase",
-  "handleAuth",
-]);
+const gv = useGameViewInject();
 
 const handleVisibleUpdate = (value) => {
-  emit("update:visible", value);
+  gv.showAuthModal.value = value;
   if (!value) {
-    emit("update:authPassphrase", "");
+    gv.authPassphrase.value = "";
   }
 };
-const handlePassphraseUpdate = (value) => emit("update:authPassphrase", value);
-const handleSubmit = () => emit("handleAuth");
 </script>
 
 <template>
   <Dialog
-    :visible="visible"
+    :visible="gv.showAuthModal.value"
     @update:visible="handleVisibleUpdate"
     modal
     header="Profile Authentication Required"
     :style="{ width: '25rem' }"
   >
     <p class="mb-4">
-      Please authenticate profile <strong>{{ authProfileId }}</strong>
+      Please authenticate profile <strong>{{ gv.authProfileId.value }}</strong>
     </p>
 
-    <form @submit.prevent="handleSubmit">
+    <form @submit.prevent="gv.handleAuth">
       <div class="flex flex-col gap-4">
         <div class="flex flex-col gap-2">
           <label>Passphrase</label>
           <Password
-            :modelValue="authPassphrase"
-            @update:modelValue="handlePassphraseUpdate"
+            v-model="gv.authPassphrase.value"
             required
             :feedback="false"
             toggleMask
@@ -54,7 +40,7 @@ const handleSubmit = () => emit("handleAuth");
           />
         </div>
 
-        <Message v-if="authError" severity="error">{{ authError }}</Message>
+        <Message v-if="gv.authError.value" severity="error">{{ gv.authError.value }}</Message>
 
         <div class="flex justify-end gap-2">
           <Button
@@ -65,8 +51,8 @@ const handleSubmit = () => emit("handleAuth");
           />
           <Button
             type="submit"
-            :label="authLoading ? 'Authenticating...' : 'Authenticate'"
-            :loading="authLoading"
+            :label="gv.authLoading.value ? 'Authenticating...' : 'Authenticate'"
+            :loading="gv.authLoading.value"
           />
         </div>
       </div>
