+++ Dashboard/src/components/game/GameViewContainer.vue
@@ -1,4 +1,5 @@
 <script setup>
+import { useGameViewInject } from "../../composables/game/injectKey";
 import Tabs from "primevue/tabs";
 import TabList from "primevue/tablist";
 import Tab from "primevue/tab";
@@ -15,133 +16,23 @@ import AuthModal from "./AuthModal.vue";
 import StatEditModal from "./StatEditModal.vue";
 import PortraitConfigModal from "./PortraitConfigModal.vue";
 
-const props = defineProps({
-  title: String,
-  activeTab: String,
-  tabs: Array,
-  config: Object,
+const gv = useGameViewInject();
 
-  profileId: Number,
-  server: String,
-  characterName: String,
-  floor: Number,
-  filteredCharacters: Array,
-  loading: Boolean,
-  error: String,
-  resultImages: Object,
-
-  canManage: Boolean,
-  manageLoading: Boolean,
-  manageError: String,
-  newCharacterName: String,
-  manageSearchQuery: String,
-  showOnlyMissingAscension: Boolean,
-  manageCharacterItems: Array,
-  filteredManageCharacters: Array,
-  hasStatEdit: Boolean,
-
-  aliasSearchQuery: String,
-  aliases: Array,
-  filteredAliases: Array,
-  showAddAliasModal: Boolean,
-  newAliasCharacter: String,
-  newAliasList: String,
-  addAliasLoading: Boolean,
-  isEditingAlias: Boolean,
-
-  hasCodesManagement: Boolean,
-  codes: Array,
-  filteredCodes: Array,
-  selectedCodes: Array,
-  newCodesInput: String,
-  codesSearchQuery: String,
-  codesLoading: Boolean,
-
-  showAuthModal: Boolean,
-  authProfileId: [String, Number],
-  authPassphrase: String,
-  authLoading: Boolean,
-  authError: String,
-
-  showEditStatModal: Boolean,
-  editStatCharacter: String,
-  editStatBase: [Number, null],
-  editStatMax: [Number, null],
-  editStatFetching: Boolean,
-  editStatLoading: Boolean,
-
-  showPortraitConfigModal: Boolean,
-  portraitConfigCharacter: String,
-  portraitConfigServerIds: Array,
-  portraitConfigServerId: [Number, null],
-  portraitConfigOffsetX: [Number, null],
-  portraitConfigOffsetY: [Number, null],
-  portraitConfigTargetScale: [Number, null],
-  portraitConfigEnableFade: [Boolean, null],
-  portraitConfigFadeStart: [Number, null],
-  portraitConfigFetching: Boolean,
-  portraitConfigSaving: Boolean,
-});
-
-const emit = defineEmits([
-  "update:activeTab",
-  "update:profileId",
-  "update:server",
-  "update:characterName",
-  "update:floor",
-  "searchCharacter",
-  "execute",
-  "update:newCharacterName",
-  "update:manageSearchQuery",
-  "update:showOnlyMissingAscension",
-  "addCharacter",
-  "deleteCharacter",
-  "editStat",
-  "update:aliasSearchQuery",
-  "update:showAddAliasModal",
-  "update:newAliasCharacter",
-  "update:newAliasList",
-  "openAddAliasModal",
-  "openEditAliasModal",
-  "handleAliasSubmit",
-  "update:selectedCodes",
-  "update:newCodesInput",
-  "update:codesSearchQuery",
-  "confirmAddCodes",
-  "confirmDeleteCodes",
-  "update:showAuthModal",
-  "update:authPassphrase",
-  "handleAuth",
-  "update:showEditStatModal",
-  "update:editStatBase",
-  "update:editStatMax",
-  "handleStatSubmit",
-  "editPortrait",
-  "update:showPortraitConfigModal",
-  "update:portraitConfigServerId",
-  "update:portraitConfigOffsetX",
-  "update:portraitConfigOffsetY",
-  "update:portraitConfigTargetScale",
-  "update:portraitConfigEnableFade",
-  "update:portraitConfigFadeStart",
-  "handlePortraitConfigSubmit",
-]);
-
-const getTabConfig = (tabId) => props.config.tabs.find((t) => t.id === tabId);
+const getTabConfig = (tabId) => gv.config.tabs.find((t) => t.id === tabId);
 </script>
 
 <template>
   <div class="game-view">
-    <h1 class="text-4xl font-bold mb-3">{{ title }}</h1>
+    <h1 class="text-4xl font-bold mb-3">{{ gv.config.title }}</h1>
 
     <Tabs
-      :value="activeTab"
-      @update:value="(value) => emit('update:activeTab', value)"
+      :value="gv.activeTab.value"
+      @update:value="(value) => (gv.activeTab.value = value)"
       scrollable
     >
       <TabList>
         <Tab
-          v-for="tab in tabs"
+          v-for="tab in gv.tabs.value"
           :key="tab.id"
           :value="tab.id"
           class="whitespace-nowrap shrink-0"
@@ -150,110 +41,32 @@ const getTabConfig = (tabId) => props.config.tabs.find((t) => t.id === tabId);
         </Tab>
       </TabList>
       <TabPanels>
-        <TabPanel v-for="tab in tabs" :key="tab.id" :value="tab.id">
+        <TabPanel v-for="tab in gv.tabs.value" :key="tab.id" :value="tab.id">
           <ManageCharactersCard
             v-if="tab.id === 'manage'"
-            :characters="filteredManageCharacters"
-            :filteredCharacters="filteredManageCharacters"
-            :newCharacterName="newCharacterName"
-            :manageSearchQuery="manageSearchQuery"
-            :showOnlyMissingAscension="showOnlyMissingAscension"
-            :manageCharacterItems="manageCharacterItems"
-            :manageLoading="manageLoading"
-            :manageError="manageError"
-            :hasStatEdit="hasStatEdit"
-            @update:newCharacterName="
-              (value) => emit('update:newCharacterName', value)
-            "
-            @update:manageSearchQuery="
-              (value) => emit('update:manageSearchQuery', value)
-            "
-            @update:showOnlyMissingAscension="
-              (value) => emit('update:showOnlyMissingAscension', value)
-            "
-            @addCharacter="emit('addCharacter')"
-            @deleteCharacter="(name) => emit('deleteCharacter', name)"
-            @editStat="(name) => emit('editStat', name)"
-            @editPortrait="(name) => emit('editPortrait', name)"
           />
 
           <ManageAliasesCard
             v-else-if="tab.id === 'aliases'"
-            :aliases="aliases"
-            :filteredAliases="filteredAliases"
-            :aliasSearchQuery="aliasSearchQuery"
-            :manageLoading="manageLoading"
-            :showAddAliasModal="showAddAliasModal"
-            :newAliasCharacter="newAliasCharacter"
-            :newAliasList="newAliasList"
-            :addAliasLoading="addAliasLoading"
-            :isEditingAlias="isEditingAlias"
-            @update:aliasSearchQuery="
-              (value) => emit('update:aliasSearchQuery', value)
-            "
-            @update:showAddAliasModal="
-              (value) => emit('update:showAddAliasModal', value)
-            "
-            @update:newAliasCharacter="
-              (value) => emit('update:newAliasCharacter', value)
-            "
-            @update:newAliasList="(value) => emit('update:newAliasList', value)"
-            @openAddAliasModal="emit('openAddAliasModal')"
-            @openEditAliasModal="(data) => emit('openEditAliasModal', data)"
-            @handleAliasSubmit="emit('handleAliasSubmit')"
           />
 
           <ManageCodesCard
             v-else-if="tab.id === 'codes'"
-            :codes="codes"
-            :filteredCodes="filteredCodes"
-            :selectedCodes="selectedCodes"
-            :newCodesInput="newCodesInput"
-            :codesSearchQuery="codesSearchQuery"
-            :codesLoading="codesLoading"
-            @update:selectedCodes="
-              (value) => emit('update:selectedCodes', value)
-            "
-            @update:newCodesInput="
-              (value) => emit('update:newCodesInput', value)
-            "
-            @update:codesSearchQuery="
-              (value) => emit('update:codesSearchQuery', value)
-            "
-            @confirmAddCodes="emit('confirmAddCodes')"
-            @confirmDeleteCodes="(list) => emit('confirmDeleteCodes', list)"
           />
 
           <CommandCard
             v-else
-            :activeTab="activeTab"
             :tabConfig="getTabConfig(tab.id)"
-            :config="config"
-            :profileId="profileId"
-            :server="server"
-            :characterName="characterName"
-            :floor="floor"
-            :filteredCharacters="filteredCharacters"
-            :loading="loading"
-            :error="error"
-            @update:profileId="(value) => emit('update:profileId', value)"
-            @update:server="(value) => emit('update:server', value)"
-            @update:characterName="
-              (value) => emit('update:characterName', value)
-            "
-            @update:floor="(value) => emit('update:floor', value)"
-            @searchCharacter="(e) => emit('searchCharacter', e)"
-            @execute="emit('execute')"
           />
         </TabPanel>
       </TabPanels>
     </Tabs>
 
-    <div v-if="resultImages[activeTab]" class="result-container mt-4">
+    <div v-if="gv.resultImages.value[gv.activeTab.value]" class="result-container mt-4">
       <Card>
         <template #content>
           <Image
-            :src="resultImages[activeTab]"
+            :src="gv.resultImages.value[gv.activeTab.value]"
             alt="Result"
             preview
             width="100%"
@@ -262,57 +75,11 @@ const getTabConfig = (tabId) => props.config.tabs.find((t) => t.id === tabId);
       </Card>
     </div>
 
-    <AuthModal
-      :visible="showAuthModal"
-      :authProfileId="authProfileId"
-      :authPassphrase="authPassphrase"
-      :authLoading="authLoading"
-      :authError="authError"
-      @update:visible="(value) => emit('update:showAuthModal', value)"
-      @update:authPassphrase="(value) => emit('update:authPassphrase', value)"
-      @handleAuth="emit('handleAuth')"
-    />
+    <AuthModal />
 
-    <StatEditModal
-      v-if="hasStatEdit"
-      :visible="showEditStatModal"
-      :characterName="editStatCharacter"
-      :baseVal="editStatBase"
-      :maxAscVal="editStatMax"
-      :loading="editStatLoading"
-      :fetching="editStatFetching"
-      @update:visible="(value) => emit('update:showEditStatModal', value)"
-      @update:baseVal="(value) => emit('update:editStatBase', value)"
-      @update:maxAscVal="(value) => emit('update:editStatMax', value)"
-      @submit="emit('handleStatSubmit')"
-    />
+    <StatEditModal v-if="gv.config.hasStatEdit" />
 
-    <PortraitConfigModal
-      :visible="showPortraitConfigModal"
-      :characterName="portraitConfigCharacter"
-      :gameId="config.id"
-      :serverIds="portraitConfigServerIds"
-      :selectedServerId="portraitConfigServerId"
-      :loading="portraitConfigSaving"
-      :fetching="portraitConfigFetching"
-      :offsetX="portraitConfigOffsetX"
-      :offsetY="portraitConfigOffsetY"
-      :targetScale="portraitConfigTargetScale"
-      :enableGradientFade="portraitConfigEnableFade"
-      :gradientFadeStart="portraitConfigFadeStart"
-      :anchorX="config.portraitAnchorX"
-      :anchorY="config.portraitAnchorY"
-      :portraitAlignX="config.portraitAlignX"
-      :portraitAlignY="config.portraitAlignY"
-      @update:visible="(value) => emit('update:showPortraitConfigModal', value)"
-      @update:selectedServerId="(value) => emit('update:portraitConfigServerId', value)"
-      @update:offsetX="(value) => emit('update:portraitConfigOffsetX', value)"
-      @update:offsetY="(value) => emit('update:portraitConfigOffsetY', value)"
-      @update:targetScale="(value) => emit('update:portraitConfigTargetScale', value)"
-      @update:enableGradientFade="(value) => emit('update:portraitConfigEnableFade', value)"
-      @update:gradientFadeStart="(value) => emit('update:portraitConfigFadeStart', value)"
-      @submit="emit('handlePortraitConfigSubmit')"
-    />
+    <PortraitConfigModal />
   </div>
 </template>
