+++ Dashboard/src/router/index.js
@@ -8,10 +8,7 @@ import DashboardHomeView from "../views/DashboardHomeView.vue";
 import ChangePasswordView from "../views/ChangePasswordView.vue";
 import UserManagementView from "../views/UserManagementView.vue";
 import DocsManagementView from "../views/DocsManagementView.vue";
-import GenshinView from "../views/GenshinView.vue";
-import HsrView from "../views/HsrView.vue";
-import ZzzView from "../views/ZzzView.vue";
-import Hi3View from "../views/Hi3View.vue";
+import GameView from "../views/GameView.vue";
 import SeaweedFilerView from "../views/SeaweedFilerView.vue";
 import PrivacyPolicyView from "../views/PrivacyPolicyView.vue";
 import TermsOfServiceView from "../views/TermsOfServiceView.vue";
@@ -69,24 +66,15 @@ const router = createRouter({
           component: DocsManagementView,
         },
         {
-          path: "genshin",
-          name: "genshin",
-          component: GenshinView,
-        },
-        {
-          path: "hsr",
-          name: "hsr",
-          component: HsrView,
-        },
-        {
-          path: "zzz",
-          name: "zzz",
-          component: ZzzView,
-        },
-        {
-          path: "hi3",
-          name: "hi3",
-          component: Hi3View,
+          path: ":game",
+          name: "game",
+          component: GameView,
+          beforeEnter: (to) => {
+            const validGames = ["genshin", "hsr", "zzz", "hi3"];
+            if (!validGames.includes(to.params.game)) {
+              return { name: "dashboard-home" };
+            }
+          },
         },
         {
           path: "seaweed-filer",
